将resolve_data_path从config.py移到utils/path_helper.py
