+++ ui/app.py
@@ -166,7 +166,7 @@ class DAVServerApp:
 
         if enable_file:
             log_file = self.settings_service.get_setting("log_file_path", "dav_server.log")
-            from config import resolve_data_path
+            from utils.path_helper import resolve_data_path
             log_file = resolve_data_path(log_file)
             log_level = self.settings_service.get_setting("log_level", "INFO")
